add: separate new file for filter tags display
